Fix load_place.py and index.html

+++ places/models.py
@@ -7,8 +7,8 @@ class Place(models.Model):
     title = models.CharField("Название места", max_length=200, db_index=True)
     description_short = models.TextField("Краткое описание", null=True, blank=True)
     description_long = HTMLField("Полное описание", null=True, blank=True)
-    latitude = models.FloatField("Широта", null=True, blank=True)
-    longitude = models.FloatField("Долгота", null=True, blank=True)
+    latitude = models.DecimalField("Широта", max_digits=10, decimal_places=7, null=True, blank=True)
+    longitude = models.DecimalField("Долгота", max_digits=10, decimal_places=7, null=True, blank=True)
     
 
     def __str__(self):
